backend/requirements.txt
-e ../packages/legal-agent-framework

fastapi>=0.115.0
uvicorn>=0.31.1
langchain-openai>=1.0.0
langgraph>=1.0.0
copilotkit>=0.1.77
ag-ui-langgraph>=0.0.24
supabase>=2.0.0
python-dotenv>=1.0.0
pypdf>=3.16.2
python-docx>=0.2.10
Pillow>=10.1.0
python-multipart>=0.0.21
httpx>=0.28.1
slowapi>=0.1.9
PyJWT[crypto]>=2.8.0

# RAG dependencies
openai>=1.10.0
cohere>=4.47
datasets>=2.14.0
tiktoken>=0.5.0
beautifulsoup4>=4.12.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
